1220 김지민 Fix: 별점 수정 시 즉시 적용 (0.5 제거)

+++ be/src/main/java/com/example/be/dto/tour/Tour.java
@@ -22,5 +22,5 @@ public class Tour {
 
     //평점
     private Integer reviewCnt;
-    private Double rateAvg;
+    private Integer rateAvg;
 }
